Set StepFunctions v2 provider as the default provider (#8779)
