+++ tests/aws/services/stepfunctions/v2/error_handling/test_task_service_sfn.py
@@ -1,7 +1,5 @@
 import json
 
-import pytest
-
 from localstack.testing.pytest import markers
 from localstack.testing.snapshots.transformer import JsonpathTransformer, RegexTransformer
 from localstack.utils.strings import short_uid
@@ -9,15 +7,7 @@ from tests.aws.services.stepfunctions.templates.base.base_templates import BaseT
 from tests.aws.services.stepfunctions.templates.services.services_templates import (
     ServicesTemplates as ST,
 )
-from tests.aws.services.stepfunctions.utils import (
-    create,
-    create_and_record_execution,
-    is_old_provider,
-)
-
-pytestmark = pytest.mark.skipif(
-    condition=is_old_provider(), reason="Test suite for v2 provider only."
-)
+from tests.aws.services.stepfunctions.utils import create, create_and_record_execution
 
 
 @markers.snapshot.skip_snapshot_verify(
