+++ tests/aws/services/stepfunctions/v2/error_handling/test_task_service_sqs.py
@@ -12,11 +12,7 @@ from tests.aws.services.stepfunctions.templates.errorhandling.error_handling_tem
 from tests.aws.services.stepfunctions.templates.services.services_templates import (
     ServicesTemplates as ST,
 )
-from tests.aws.services.stepfunctions.utils import create_and_record_execution, is_old_provider
-
-pytestmark = pytest.mark.skipif(
-    condition=is_old_provider(), reason="Test suite for v2 provider only."
-)
+from tests.aws.services.stepfunctions.utils import create_and_record_execution
 
 
 @markers.snapshot.skip_snapshot_verify(
