+++ tests/aws/services/stepfunctions/v2/scenarios/test_base_scenarios.py
@@ -18,11 +18,6 @@ from tests.aws.services.stepfunctions.utils import (
     await_execution_terminated,
     create,
     create_and_record_execution,
-    is_old_provider,
-)
-
-pytestmark = pytest.mark.skipif(
-    condition=is_old_provider(), reason="Test suite for v2 provider only."
 )
