+++ tests/aws/services/stepfunctions/v2/services/test_events_task_service.py
@@ -8,15 +8,7 @@ from localstack.utils.strings import short_uid
 from tests.aws.services.stepfunctions.templates.services.services_templates import (
     ServicesTemplates as ST,
 )
-from tests.aws.services.stepfunctions.utils import (
-    create_and_record_execution,
-    is_old_provider,
-    record_sqs_events,
-)
-
-pytestmark = pytest.mark.skipif(
-    condition=is_old_provider(), reason="Test suite for v2 provider only."
-)
+from tests.aws.services.stepfunctions.utils import create_and_record_execution, record_sqs_events
 
 
 @markers.snapshot.skip_snapshot_verify(
