+++ tests/aws/services/stepfunctions/v2/test_sfn_api_versioning.py
@@ -11,11 +11,6 @@ from tests.aws.services.stepfunctions.utils import (
     await_execution_terminated,
     await_state_machine_version_listed,
     await_state_machine_version_not_listed,
-    is_old_provider,
-)
-
-pytestmark = pytest.mark.skipif(
-    condition=is_old_provider(), reason="Test suite for v2 provider only."
 )
