+++ tests/aws/services/stepfunctions/v2/timeouts/test_timeouts.py
@@ -13,11 +13,6 @@ from tests.aws.services.stepfunctions.templates.timeouts.timeout_templates impor
 from tests.aws.services.stepfunctions.utils import (
     await_execution_terminated,
     create_and_record_execution,
-    is_old_provider,
-)
-
-pytestmark = pytest.mark.skipif(
-    condition=is_old_provider(), reason="Test suite for v2 provider only."
 )
